pyproject.toml
[project]
name = "longevity"
version = "0.9.0"
description = "AI-платформа для извлечения данных о связывании лигандов и белков"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "g4f[all]>=0.5.7.5",
    "ipykernel>=6.29.5",
    "loguru>=0.7.3",
    "openai-agents>=0.2.2",
    "rdkit>=2025.3.3",
    "pandas>=2.0.0",
    "requests>=2.25.0",
    "tenacity>=9.1.2",
    "pypdf>=5.8.0",
    "pytesseract>=0.3.13",
    "pymupdf>=1.26.3",
    "opencv-python>=4.12.0.88",
    "rich>=14.0.0",
    "textual>=4.0.0",
    "textual-dev>=1.7.0",
]
